# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Leet-Pi API
  description: |
    Backend escalable para MVP de evaluación de ejercicios de programación en C.

    ## Autenticación
    La mayoría de los endpoints requieren un JWT Bearer token.  
    Obtené el token haciendo `POST /api/auth/login` con un token Auth0 válido en el header `Authorization: Bearer <auth0_token>`.  
    El servidor retorna el JWT de la plataforma en el header de respuesta **`X-Auth-Token`**. Usá ese valor para llamadas autenticadas subsecuentes.

    ## Roles
    - **student** – puede enviar soluciones y ver sus propias submissions.
    - **admin** – acceso a operaciones administrativas de guías, ejercicios y usuarios normales.
    - **superadmin** – nivel más alto: todo lo que puede admin más gestión de admins y asignación de roles.
  version: 1.0.0
  contact:
    name: ITBA – leet-pi team
servers:
  - url: http://localhost:3001/api
    description: Desarrollo local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por la plataforma (`X-Auth-Token`). Obtenido vía `POST /api/auth/login`.
    Auth0Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT emitido directamente por Auth0. Requerido únicamente en `POST /api/auth/login`.

  # ─── Common response schemas ───────────────
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          description: Payload variado según el endpoint
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            statusCode:
              type: integer
          required:
            - message
            - statusCode
      required:
        - success
        - error

    # ─── Domain schemas ──────────────────────
    AdminUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        enabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
            enum: [admin, superadmin]
      required: [id, email, firstName, lastName, enabled, roles]
...
paths:
...
  /admin/users/{userId}/enabled:
    patch:
      tags: [Admin]
      summary: Habilitar o deshabilitar un usuario
      description: |
        Cambia el estado `enabled` de un usuario.
        **Reglas de acceso:**
        - Nadie puede modificar un `superadmin`.
        - Solo un `superadmin` puede modificar a un `admin`.
        - Un `admin` puede modificar usuarios normales.
        - Nadie puede modificar su propio estado.
      operationId: updateUserEnabled
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
            example:
              enabled: false
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminUser'
              example:
                success: true
                message: User updated successfully
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  email: user@example.com
                  firstName: Juan
                  lastName: Pérez
                  enabled: false
                  roles: []
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/users/upload-csv:
    post:
      tags: [Admin]
      summary: Cargar CSV de usuarios
      description: |
        Sube un archivo CSV para habilitar/deshabilitar usuarios en masa.  
        El CSV debe contener una columna con los emails de los usuarios habilitados.
        - Usuarios en el CSV → se habilitan (se crean si no existen).
        - Usuarios en la DB que **no** están en el CSV → se deshabilitan.

        Límite de archivo: **10 MB**.
      operationId: uploadCSV
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [csv]
              properties:
                csv:
                  type: string
                  format: binary
                  description: Archivo CSV (máx. 10 MB)
      responses:
        '200':
          description: CSV procesado sin errores
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          enabled:
                            type: integer
                          disabled:
                            type: integer
                          created:
                            type: integer
                          totalProcessed:
                            type: integer
                          errors:
                            type: array
                            items:
                              type: string
              example:
                success: true
                message: CSV processed successfully
                data:
                  enabled: 50
                  disabled: 10
                  created: 5
                  totalProcessed: 60
        '207':
          description: CSV procesado con errores parciales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Archivo inválido o CSV vacío
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/users/status:
    get:
      tags: [Admin]
      summary: Estadísticas de usuarios
      description: Retorna el total, habilitados y deshabilitados de usuarios en la plataforma.
      operationId: getUserStatus
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Estadísticas obtenidas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total:
                            type: integer
                          enabled:
                            type: integer
                          disabled:
                            type: integer
              example:
                success: true
                message: User status retrieved successfully
                data:
                  total: 120
                  enabled: 100
                  disabled: 20
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/users/{userId}/roles:
    put:
      tags: [Admin]
      summary: Actualizar roles de un usuario (solo superadmin)
      description: |
        Reemplaza los roles del usuario indicado. **Requiere rol `superadmin`.**
        **Restricciones:**
        - No se puede modificar los propios roles.
        - No se puede modificar los roles de un `superadmin`.
        - Solo se puede asignar/quitar el rol `admin` (no `superadmin`).

        Roles válidos para asignar: `admin`.
      operationId: updateUserRoles
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              properties:
                roles:
                  type: array
                  items:
                    type: string
                    enum: [admin]
                  description: Lista de roles a asignar (reemplaza los existentes). Solo se admite 'admin'.
            example:
              roles: [student]
      responses:
        '200':
          description: Roles actualizados
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userId:
                            type: string
                          roles:
                            type: array
                            items:
                              type: string
              example:
                success: true
                message: User roles updated successfully
                data:
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  roles: [student]
        '400':
          description: Roles inválidos o request mal formado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/guides/{guideNumber}:
    patch:
      tags: [Admin]
      summary: Actualizar guía
      description: |
        Actualiza el estado (enabled) y/o el deadline de una guía.  
        Al menos uno de `enabled` o `deadline` debe estar presente.
      operationId: updateGuide
      security:
        - BearerAuth: []
      parameters:
        - name: guideNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: Número de la guía
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                  description: Habilitar o deshabilitar la guía
                deadline:
                  type: string
                  format: date-time
                  description: Fecha límite de entrega (ISO 8601). Null para quitar deadline.
            example:
              enabled: true
              deadline: '2024-06-30T23:59:00.000Z'
      responses:
        '200':
          description: Guía actualizada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          guideNumber:
                            type: integer
                          enabled:
                            type: boolean
                          deadline:
                            type: string
                            format: date-time
                            nullable: true
              example:
                success: true
                message: Guide updated successfully
                data:
                  guideNumber: 1
                  enabled: true
                  deadline: '2024-06-30T23:59:00.000Z'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/guides/{guideNumber}/exercises/{exerciseNumber}:
    patch:
      tags: [Admin]
      summary: Actualizar ejercicio
      description: Habilita o deshabilita un ejercicio específico dentro de una guía.
      operationId: updateExercise
      security:
        - BearerAuth: []
      parameters:
        - name: guideNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: Número de la guía
        - name: exerciseNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: Número del ejercicio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
                  description: Nueva disponibilidad del ejercicio
            example:
              enabled: false
      responses:
        '200':
          description: Ejercicio actualizado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          guideNumber:
                            type: integer
                          exerciseNumber:
                            type: integer
                          enabled:
                            type: boolean
              example:
                success: true
                message: Exercise updated successfully
                data:
                  guideNumber: 1
                  exerciseNumber: 3
                  enabled: false
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'